@fafhrd91 2020-08-09
I am not really sure, but your change should affect only "fortunes" test,
and it is lower after the change, before 5_002_304 vs 3_103_472 after. and other results should not change.

could you re-run bench for "fortunes" test only?