@pickfire 2020-08-09
I got
```
--------------------------------------------------------------------------------
Benchmarking ntex
--------------------------------------------------------------------------------
ntex: Traceback (most recent call last):
ntex:   File "/FrameworkBenchmarks/toolset/benchmark/benchmarker.py", line 167, in __run_test
ntex:     accepting_requests = test.is_accepting_requests()
ntex:   File "/FrameworkBenchmarks/toolset/benchmark/framework_test.py", line 77, in is_accepting_requests
ntex:     self.runTests[test_type].get_url())
ntex: KeyError: None
ntex: Error during test: ntex
ntex: Total test time: 1m 2s
tfb: Total time building so far: 1m 2s
tfb: Total time verifying so far: 0s
tfb: Total execution time so far: 1m 11s
```
with `./tfb --test ntex --type fortune`, seemed like there are no results, did I do anything wrong?