@pickfire 2020-08-10
@fafhrd91 By the way, do you know how drogon getFastDbClient is able to be have database calls so fast? I saw the numbers for database updates was very impressive for the last round.