@fafhrd91 2020-08-10
i think it uses batch mode on libpq level. need to investigate and implement the same