@pickfire 2020-08-17
How come tfb-status.techempower.com gives `hello, world `? Will the results be out?