@nbrady-techempower 2020-08-17
@pickfire tfb-status went down over the weekend due to an outage. we're getting it back up today. thanks.