@fafhrd91 2020-09-01
@pickfire performance is worse than before. change should be reverted